파일명, 코드 정리
