+++ src/hooks/useMoveTile.tsx
@@ -1,11 +1,5 @@
 import { addKeyObserver, removeKeyObserver } from '@utils/keyboard';
-import {
-  isGameOver,
-  moveDown,
-  moveLeft,
-  moveRight,
-  moveUp,
-} from '@utils/tile2';
+import { isGameOver, moveDown, moveLeft, moveRight, moveUp } from '@utils/tile';
 import { useCallback, useEffect } from 'react';
 
 export default function useMoveTile({
